pyproject.toml
[tool.poetry]
name = "soft-sensor-monitor"
version = "0.1.0"
description = ""
authors = ["Peter Pfaffeneder <peter.pfaffeneder@gmail.com>"]
readme = "README.md"
packages = [{include = "src"}]

[tool.poetry.dependencies]
python = ">=3.10,<3.12"
matplotlib = "^3.7.1"
numpy = "^1.24.2"
pandas = "^1.5.3"
scikit-learn = "^1.2.2"
scikit-image = "^0.20.0"
pyserial = "^3.5"
ruff = "^0.0.269"

[tool.poetry.dev-dependencies]
pytest = "^7.2.2"
pre-commit = "^3.2.2"
mypy = "^1.2.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
pythonpath = [
  "src"
]
minversion = "6.0"
addopts = "-ra -q"
testpaths = [
  "src/app/",
  "tests/"
]

[tool.ruff]
line-length = 150
